requirements.txt
# openai library
openai

# http client
requests

# simply ui framework
streamlit

# .env utility
python-dotenv

# jupyter notbook
jupyter

# pdf library
pypdf2
pdfplumber
pymupdf
camelot-py[base]
pandas
tabulate
pyhwp
